resource: notes on Env
went through Min's `Env.hs` --- mostly his code, with lots of comments
which helped me understand it, particularly towards the end, as well as
notes on how I would like to change it to suit my own needs better.

`Env.hs` serves as a resource to understand the implementation of such
an environment later (soon).